change veterinarian to place, more general
